use a static link with libcolm.a when running out of the source tree
A dynamic link is proving complicated on OSX. When the package is configured to
install to a new install location, but make install hasn't happened yet the
link fails This may indicate we are not getting the code we think we are. We
can eschew all these risks with a static link and reserve dynamic linking for
when we are running from the installed location.

This patch also includes some testing fixes