first cut of a redefine statement
Redefine happens during declaration. The existing langEl is removed from the
map and a newly-declared lel replaces it. If the new list of productions
references [...] the productions from the previous lel are moved into the new
one, in its place. A copy would be preferrable, as move may require some fixup,
but move gives an initial functional implementation. refs #112