use fopencookie to avoid leaking FILE structs
If we use fdopen for stdin/out/err we cannot close it to free the file struct
without also closing the file descriptor. If fopencookie is available, use that
to wrap the file descriptor, but allow closing without closing the fd. This is
useful when embedding in long running programs.