casting template type pointers to void* before passing to memmove funcs
More recent compilers will now warn about moving and copying non-trivial types,
due to the potential for internal pointers. This is a risk the libraries always
had, and it is expected to not put those things into the containers. Casting to
void* removes the new warnings and better expresses what the code is doing.