fix vm_pop_type for big-endian systems
Change the pop to cast the pointer and take the value using a pointer to the
requested type, then return that tmp.

Previously we were taking the value using the tree type, then casting the
value. During the cast of the value the high bits are dropped, and that's where
the value actually lives, when on a big-endian system. Didn't notice the
problem on little-endian systems because it happens to work fine there.

Fixes #126.