if buildDir is not set with -B, try to find a default
When running from the build directory, if one forgets to give -B, and colm was
previously installed at the current prefix, colm programs will build against
the installed version, not the local source tree. This won't be immediately
obvious and could lead to confusion during development.

To prevent this, try to find the buildDir if one is not given. The tests can
continue to specify the buildDir with -B, which is the most safe way.

The buildDir is defaulted by testing the buildDir known at compile time against
LD_LIBRARY_PATH. Libtool on ubuntu 20.04 will place the buildDir at the head of
LD_LIBRARY_PATH. It seems likely that it would do this on all systems, making
it a seemingly good test.