Use /proc/self/exe to detect colm build directory
The canonical identity of a file on Unix is the combination of device
and inode numbers. This code checks both possible paths for the
colm executable: directly in the build directory (when colm is linked
against libcolm statically) or in libtool's `objdir` subdirectory (when
colm is linked against libcolm dynamically).

This fails in two situations I can see: when /proc is not mounted or
doesn't exist, or when colm is installed using a hard link.